backend/requirements.txt
# FastAPI Backend Dependencies
# Python Version: 3.11 or 3.12 recommended (3.14 not yet fully supported by all packages)
#
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database & Vector Store
pymongo>=4.6.0,<5.0.0
motor>=3.3.0,<4.0.0
# Using 0.4.x for better Windows compatibility (avoids onnxruntime issues)
chromadb==0.4.24
# Pin numpy < 2.0 for ChromaDB 0.4.24 compatibility
numpy>=1.24.0,<2.0.0

# LangChain & LangGraph (Updated for LangGraph-first architecture)
# Using conservative versions for compatibility
langchain>=0.1.0,<0.3.0
langchain-core>=0.1.0,<0.3.0
langchain-openai>=0.0.5,<0.2.0
langgraph>=0.0.50
langchain-community>=0.0.16,<0.3.0

# Embeddings (simplified - no specific torch version)
sentence-transformers>=2.3.0

# Authentication & Security
pyjwt>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0,<5.0.0
python-jose[cryptography]>=3.3.0

# Environment & Config
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Utilities
httpx>=0.26.0
aiofiles>=23.2.0
python-dateutil>=2.8.0

# Document Processing
PyPDF2>=3.0.0
# python-docx>=0.8.11  # Uncomment when DOCX support is needed

# Image Processing & OCR
Pillow>=10.0.0
easyocr>=1.7.0
opencv-python-headless>=4.8.0,<4.10.0  # Compatible with numpy 1.x

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
